refactor: 🧩 Move fetchers to collections
