+++ src/collections/collections.js
@@ -1,7 +1,7 @@
 export * as localizations from "./localizations.js";
 export * as asserts from "./asserts.js";
 export * as fetcherTemplates from "../utils/Fetcher/fetcherTemplates.js";
-export * as fetchers from "../utils/fetchers/fetchers.js";
+export * as fetchers from "./fetchers/fetchers.js";
 export * as fetcherSenders from "../utils/Fetcher/Fetcher.js";
 export * as modes from "../utils/modes.js";
 export * as queries from "../utils/querys.js";
