+++ src/collections/fetchers/jikan.js
@@ -1,6 +1,4 @@
-import { fetcherTemplates, queries } from "../../collections/collections.js";
-import { asserts } from "../../collections/collections.js";
-import { localizations } from "../../collections/collections.js";
+import { asserts, fetcherTemplates, localizations, queries } from "../collections.js";
 
 export const getMediaById = (type, id) => {
   asserts.assertTrue(id);
